feat(settings): Money section combines Vendors / Expense Categories / QuickBooks / Stripe (#230)
Slice 4 of the Settings redesign (PRD #226):

- /settings/money: new combined section with four tabs (Vendors, Expense
  Categories, QuickBooks, Stripe). The four existing page bodies are
  extracted into co-located *-tab.tsx components; UI and behavior are
  unchanged inside each tab.
- Server-side redirects from /settings/{vendors,expense-categories,
  accounting,stripe} to /settings/money?tab=<key> so bookmarks and
  in-app links keep working.
- Renames the PRD calls out: "Accounting" → "QuickBooks" (it's only QB),
  "Stripe Payments" → "Stripe" (it's only Stripe; the Payments suffix
  collided with the unrelated /settings/payments page).
- The QuickBooks tab is a server component that does the same
  admin / access_settings / manage_accounting gate the old
  /settings/accounting page did. When the user lacks access the tab body
  shows a no-access notice instead of redirecting, so the other three
  Money tabs remain accessible.
- The Stripe tab is a server component that does the same
  loadStripeConnection + stripe_events lookup the old /settings/stripe
  page did and passes the result into StripeSettingsClient.
- QB and Stripe OAuth callbacks (/api/qb/callback, /api/stripe/connect/
  callback) now redirect directly to /settings/money?tab=... instead of
  the old URL → redirect → new URL two-hop, so query params like
  ?oauth_error=... and ?connected=1 survive.
- Inbound links from qb-sync-tab, qb-expired-banner, and the QB setup
  wizard updated to /settings/money?tab=quickbooks.
- settings-nav.ts: one new Money entry (CircleDollarSign icon); four old
  entries removed. Mobile dropdown picks this up automatically.
