+++ src/lib/settings-nav.ts
@@ -2,16 +2,13 @@ import {
   Building2,
   Palette,
   ListChecks,
-  Store,
-  Receipt,
   Users,
   Mail,
   FileSignature,
   FileText,
   Download,
   Send,
-  Link2,
-  CreditCard,
+  CircleDollarSign,
   Library,
   LayoutTemplate,
 } from "lucide-react";
@@ -35,12 +32,7 @@ export const settingsNavItems: SettingsNavItem[] = [
     icon: LayoutTemplate,
   },
   { href: "/settings/pdf-presets", label: "PDF Presets", icon: FileText },
-  { href: "/settings/vendors", label: "Vendors", icon: Store },
-  {
-    href: "/settings/expense-categories",
-    label: "Expense Categories",
-    icon: Receipt,
-  },
+  { href: "/settings/money", label: "Money", icon: CircleDollarSign },
   { href: "/settings/people", label: "People", icon: Users },
   { href: "/settings/email", label: "Email Accounts", icon: Mail },
   {
@@ -54,8 +46,6 @@ export const settingsNavItems: SettingsNavItem[] = [
     icon: FileText,
   },
   { href: "/settings/contracts", label: "Contracts", icon: Send },
-  { href: "/settings/accounting", label: "Accounting", icon: Link2 },
-  { href: "/settings/stripe", label: "Stripe Payments", icon: CreditCard },
   { href: "/settings/payments", label: "Outgoing Emails", icon: Mail },
   { href: "/settings/reports", label: "Reports", icon: FileText },
   { href: "/settings/data", label: "Data", icon: Download },
